feat: move classic poll command and migrate existing poll logic
